refactor: move manageGroupSubscriptionsFilter inside apim-filters
