+++ packages/external-clients/src/apim-client/__tests__/apim-filters.test.ts
@@ -1,5 +1,5 @@
-import { describe, it, expect } from "vitest";
 import * as O from "fp-ts/lib/Option";
+import { describe, expect, it } from "vitest";
 import {
   ApimFilterType,
   buildApimFilter,
@@ -7,6 +7,7 @@ import {
   FilterFieldEnum,
   FilterSupportedFunctionsEnum,
   FilterSupportedOperatorsEnum,
+  manageGroupSubscriptionsFilter,
   subscriptionsByIdsApimFilter,
   subscriptionsExceptManageOneApimFilter,
 } from "../apim-filters";
@@ -171,5 +172,39 @@ describe("subscriptionsExceptManageOneApimFilter", () => {
 
     // then
     expect(filter).toStrictEqual("not(startswith(name, 'MANAGE-'))");
+  });
+});
+
+describe("manageGroupSubscriptionsFilter", () => {
+  it("should return a MANAGE-GROUP- filter when no groups are provided", () => {
+    const res = manageGroupSubscriptionsFilter();
+
+    expect(res).toEqual("startswith(name, 'MANAGE-GROUP-')");
+  });
+
+  it("should return an empty filter when empty array is provided", () => {
+    const res = manageGroupSubscriptionsFilter([]);
+
+    expect(res).toEqual("");
+  });
+
+  it("should return a single MANAGE-GROUP- filter when a single group id provided", () => {
+    const groupId = "g1";
+    const res = manageGroupSubscriptionsFilter([groupId]);
+
+    expect(res).toEqual(`name eq 'MANAGE-GROUP-${groupId}'`);
+  });
+
+  it("should return a concat MANAGE-GROUP- filters based on the group ids provided", () => {
+    const groupids = new Array(3);
+    for (let index = 0; index < groupids.length; index++) {
+      groupids[index] = "id" + index;
+    }
+    const res = manageGroupSubscriptionsFilter(groupids);
+
+    const expectedRes = groupids
+      .map((groupId) => `name eq 'MANAGE-GROUP-${groupId}'`)
+      .join(" or ");
+    expect(res).toEqual(expectedRes);
   });
 });
